Update xresnet2.py
The bottleneck of the code is in the _make_layer function. This function is responsible for creating the residual blocks that make up the XResNet model. The bottleneck occurs in the line where the downsample variable is initialized. This variable is used to create a shortcut connection that bypasses the residual block if the input and output dimensions are not the same. However, the downsample variable is always initialized to a sequential container of two layers, even if the input and output dimensions are the same. This means that the shortcut connection is always created, even if it is not necessary. This can cause a bottleneck in the model, as the shortcut connection will add unnecessary computational overhead.

To fix this bottleneck, we can simply check if the input and output dimensions are the same before initializing the downsample variable. If they are the same, we can simply set the downsample variable to None. This will prevent the shortcut connection from being created, and will improve the performance of the model.